+++ src/main/java/chen/binflare/dao/UserCredentialDAO.java
@@ -2,5 +2,5 @@ package chen.binflare.dao;
 
 import chen.binflare.entity.UserCredentialDO;
 
-public interface UserCredentialDAO extends BaseRepository<UserCredentialDO, Long> {
+public interface UserCredentialDAO extends BaseRepository<UserCredentialDO, Integer> {
 }
