+++ src/main/java/chen/binflare/dao/UserDAO.java
@@ -2,5 +2,5 @@ package chen.binflare.dao;
 
 import chen.binflare.entity.UserDO;
 
-public interface UserDAO extends BaseRepository<UserDO, Long> {
+public interface UserDAO extends BaseRepository<UserDO, Integer> {
 }
